backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]>=3.2.10
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.1

# File processing
Pillow==11.0.0
opencv-python-headless==4.10.0.84
pyzbar==0.1.9
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3

# XML parsing for SAFT
lxml==4.9.3
xmlschema==2.5.0

# Storage
boto3==1.29.7

# Excel parsing
openpyxl==3.1.2

# Utils
python-dateutil==2.8.2
requests>=2.31.0
